Updated userInputModel

+++ backend/src/main/java/gs/repository/UserRepo.java
@@ -4,11 +4,10 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
 
 import gs.model.user.User;
 
-public interface UserRepo extends JpaRepository<User, Integer>{
+public interface UserRepo extends JpaRepository<User, Long>{
 
     @Query(value = "select * from user;", nativeQuery = true)
     List<Object[]> getUser();
